% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_config.R
\name{AND.multiple}
\alias{AND.multiple}
\title{AND.multiple}
\usage{
AND.multiple(...)
}
\arguments{
\item{...}{Multiple filtering conditions}
}
\value{
A single condition reduced
}
\description{
An internal function for handling multiple filtering conditions.
Reduce conditions with "&" operations.
}
\keyword{internal}
