% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_annotation.R
\name{AnnotateVCFObj}
\alias{AnnotateVCFObj}
\title{AnnotateVCFObj}
\usage{
AnnotateVCFObj(vcf.obj, df.annotation.db, columns.to.include,
  include.all.columns = FALSE)
}
\arguments{
\item{vcf.obj}{\code{\link{ReadVCF}}}

\item{df.annotation.db}{A data.frame from \code{\link{PrepareAnnotationDB}}.
This data.frame must have the columns 'CHROM', 'POS', 'REF', 'ALT'}

\item{columns.to.include}{A character vector of columns to include.
Note that existing columns in vcf.obj will not be affected.}

\item{include.all.columns}{A boolean value. If TRUE, then annotates vcf.obj with
all columns present in df.variants.of.interest. If FALSE, columns.to.include must be
supplied.}
}
\value{
An annotated vcf.obj
}
\description{
Annotates a vcf.obj using df.variants.of.interest (from \code{\link{PrepareAnnotationDB}})
}
