% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_optimization.R
\name{CheckIfVariantRefinementIsNecessary}
\alias{CheckIfVariantRefinementIsNecessary}
\title{CheckIfVariantRefinementIsNecessary}
\usage{
CheckIfVariantRefinementIsNecessary(vcf.obj, bsg, df.mut.pat.ref.sigs,
  target.mut.sigs, sequencing.artifact.mut.sigs,
  init.artifact.stop = 0.05, verbose = TRUE)
}
\arguments{
\item{vcf.obj}{A list from ReadVCF}

\item{bsg}{BSgenome.Hsapiens.UCSC object}

\item{df.mut.pat.ref.sigs}{A data.frame from MutPatParseRefMutSigs}

\item{target.mut.sigs}{A character vector of target mutational signatures from reference mutational signatures.}

\item{sequencing.artifact.mut.sigs}{A character vector of sequencing artifact mutational signatures from reference mutational signatures.}

\item{init.artifact.stop}{Numeric value less than 1. If the sum of sequencing artifact weights in vcf.obj is less than or equal to this value then
this function returns judgment = FALSE, otherwise returns judgment = TRUE.}

\item{verbose}{If TRUE, provides process detail. Default value is TRUE.}
}
\value{
A list with the following elements
\itemize{
 \item{judgment}{A boolean value}
 \item{seq.art.sigs.weights.sum}{A numeric value. Sum of sequencing artifact weights.}
}
}
\description{
Checks if variant refinement is necessary by identifying mutational signatures
related to sequencing artifact in the vcf.obj (set of original unrefined point mutations).
}
