% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_constants.R
\docType{data}
\name{Chromosome.Names}
\alias{Chromosome.Names}
\title{Constant}
\format{An object of class \code{character} of length 25.}
\usage{
Chromosome.Names
}
\description{
Chromosome names for FIREVAT.
Chromosome names should be given in the format of "chr" + chromosome number.
}
\keyword{datasets}
