% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_maths.R
\name{ComputeZScore}
\alias{ComputeZScore}
\title{ComputeZScore}
\usage{
ComputeZScore(values, x)
}
\arguments{
\item{values}{a numeric vector}

\item{x}{a numeric value}
}
\value{
a numeric value corresponding to the z-score of x
}
\description{
Returns a z-score of x given a distribution of values
}
