% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_maths.R
\name{ComputeZScoreEquiValue}
\alias{ComputeZScoreEquiValue}
\title{ComputeZScoreEquiValue}
\usage{
ComputeZScoreEquiValue(z.score, values)
}
\arguments{
\item{z.score}{numeric value}

\item{values}{numeric vector}
}
\value{
a numeric value corresponding to the specified z.score in the 'values' distribution
}
\description{
Returns a numeric value that is equivalent to the specified z.score
  in the distribution of 'values'
}
