% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_maths.R
\name{DecimalCeiling}
\alias{DecimalCeiling}
\title{DecimalCeiling}
\usage{
DecimalCeiling(value, decimal)
}
\arguments{
\item{value}{numeric value (decimal)}

\item{decimal}{numeric value (e.g. 0.1, 0.001)}
}
\value{
a numeric value
}
\description{
Returns the ceiling of a decimal value
  e.g. value = 0.15, decimal = 0.1 returns 0.2
}
