% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_filter.R
\name{DefaultFilterToBinary}
\alias{DefaultFilterToBinary}
\title{Transform default filtering parameters to a binary vector}
\usage{
DefaultFilterToBinary(vcf.filter, params.bit.len)
}
\arguments{
\item{vcf.filter}{A list generated in \code{\link{MakeFilter}}}

\item{params.bit.len}{A list with bit lengths of filtering parameters which is generated from \code{\link{ParameterToBits}}}
}
\value{
A binary vector
}
\description{
This function transforms default filtering parameter to binary vector
which can be used as a suggested solution in GA algorithm.
}
