% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_utils.R
\name{EnumerateTriNucCounts}
\alias{EnumerateTriNucCounts}
\title{EnumerateTriNucCounts}
\usage{
EnumerateTriNucCounts(spectrum)
}
\arguments{
\item{spectrum}{a numeric vector with 96 numeric values}
}
\value{
a numeric vector of length 6 corresponding to the counts of each trinucleotide change (C>A, C>G, C>T, T>A, T>C, T>G)
}
\description{
Returns C>A, C>G, C>T, T>A, T>C, T>G counts
}
