% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_objective_functions.R
\name{Exp.Weighted.A.Ref.Obj.Fn}
\alias{Exp.Weighted.A.Ref.Obj.Fn}
\title{Exp.Weighted.A.Ref.Obj.Fn}
\usage{
Exp.Weighted.A.Ref.Obj.Fn(C.refined, A.refined, C.artifactual,
  A.artifactual)
}
\arguments{
\item{C.refined}{A numeric value between 0 and 1.}

\item{A.refined}{A numeric value between 0 and 1.}

\item{C.artifactual}{A numeric value between 0 and 1.}

\item{A.artifactual}{A numeric value between 0 and 1.}
}
\value{
A numeric value between 0 and 1.
}
\description{
Exponentially weighted objective function
}
