% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_objective_functions.R
\name{Exp.Weighted.Obj.Fn.1}
\alias{Exp.Weighted.Obj.Fn.1}
\title{Exp.Weighted.Obj.Fn.1}
\usage{
Exp.Weighted.Obj.Fn.1(C.refined, A.refined, C.artifactual, A.artifactual)
}
\arguments{
\item{C.refined}{A numeric value between 0 and 1.}

\item{A.refined}{A numeric value between 0 and 1.}

\item{C.artifactual}{A numeric value between 0 and 1.}

\item{A.artifactual}{A numeric value between 0 and 1.}
}
\value{
A numeric value between 0 and 1.
}
\description{
Calculates the exponentially weighted objective value
for FIREVAT GA optimization.
}
