% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_objective_functions.R
\name{Exp.Weighted.Refined.Seq.Art.Only.Obj.Fn}
\alias{Exp.Weighted.Refined.Seq.Art.Only.Obj.Fn}
\title{Exp.Weighted.Refined.Seq.Art.Only.Obj.Fn}
\usage{
Exp.Weighted.Refined.Seq.Art.Only.Obj.Fn(C.refined, A.refined,
  C.artifactual, A.artifactual)
}
\arguments{
\item{C.refined}{A numeric value between 0 and 1.}

\item{A.refined}{A numeric value between 0 and 1.}

\item{C.artifactual}{A numeric value between 0 and 1.}

\item{A.artifactual}{A numeric value between 0 and 1.}
}
\value{
A numeric value between 0 and 1.
}
\description{
Calculates the Euclidean-distance of logarithmically weighted
objective value for FIREVAT GA optimization.
}
