% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_vcf.R
\name{FilterByStrandBiasAnalysis}
\alias{FilterByStrandBiasAnalysis}
\title{FilterByStrandBiasAnalysis}
\usage{
FilterByStrandBiasAnalysis(refined.vcf.obj, artifactual.vcf.obj,
  perform.fdr.correction, filter.by.strand.bias.analysis.cutoff)
}
\arguments{
\item{refined.vcf.obj}{A list of vcf data}

\item{artifactual.vcf.obj}{A list of vcf data}

\item{perform.fdr.correction}{A boolean value.}

\item{filter.by.strand.bias.analysis.cutoff}{A numeric value.}
}
\value{
A list with filtering parameter values
\itemize{
 \item{refined.vcf.obj}{ updated refined.vcf.obj}
 \item{artifactual.vcf.obj}{ updated artifactual.vcf.obj}
}
}
\description{
Filters refined.vcf.obj by strand bias analysis and
moves these filtered variants to artifactual.vcf.obj
}
