% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_filter.R
\name{FilterVCF}
\alias{FilterVCF}
\title{FilterVCF}
\usage{
FilterVCF(vcf.obj, vcf.filter, config.obj, include.array = NULL,
  force.include = FALSE, verbose = TRUE)
}
\arguments{
\item{vcf.obj}{A list from ReadVCF}

\item{vcf.filter}{A list from MakeMuTect2Filter}

\item{config.obj}{A list from ParseConfigFile}

\item{include.array}{A boolean vector}

\item{force.include}{A boolean value. If TRUE, then uses 'include.array'}

\item{verbose}{If true, provides process detail}
}
\value{
A list with the following elements
\itemize{
  \item{1) Mutations which passed filtering}{vcf.obj.filtered = vcf.obj (list with data, header, genome)}
  \item{2) Mutations which did not pass filtering}{vcf.obj.artifact = vcf.obj (list with data, header, genome)}
}
}
\description{
Filter vcf based on the filter
Filtering parameters are saved in config.obj
Split vcf.obj into vcf.obj.filtered & vcf.obj.artifact based on vcf.filter
}
