% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_optimization.R
\name{GADecodeBinaryString}
\alias{GADecodeBinaryString}
\title{GADecodeBinaryString}
\usage{
GADecodeBinaryString(string, data)
}
\arguments{
\item{string}{A binary expression of parameters}

\item{data}{A list which contains FIREVAT data}
}
\value{
A numeric vector corresponding to decoded string representation
of parameters
}
\description{
This function returns decimal vector decoded from binary string.
}
\keyword{internal}
