% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_optimization.R
\name{GAMonitorFn}
\alias{GAMonitorFn}
\title{GAMonitorFn}
\usage{
GAMonitorFn(obj, data)
}
\arguments{
\item{obj}{= A factor returned by GA package}

\item{data}{A list with the following data
\itemize{
 \item{"vcf.obj"}{\code{\link{ReadVCF}}}
 \item{"vcf.filter"}{a list}
 \item{"vcf.calling.method"}{string value}
 \item{"df.ref.mut.sig"}{a data.frame}
 \item{"target.mut.sigs"}{a character vector}
 \item{"sequencing.artifact.mut.sigs"}{a charcter vector}
 \item{"config.obj"}{\code{\link{ParseConfigFile}}}
}}
}
\description{
Prints GA optimization iteration to console
}
\keyword{internal}
