% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_optimization.R
\name{GAOptimizationObjFn}
\alias{GAOptimizationObjFn}
\title{GAOptimizationObjFn}
\usage{
GAOptimizationObjFn(params.x, data)
}
\arguments{
\item{data}{= A list with the following data
\itemize{
 \item{"vcf.obj"}{\code{\link{ReadVCF}}}
 \item{"vcf.filter"}{a list}
 \item{"vcf.calling.method"}{string value}
 \item{"df.ref.mut.sig"}{a data.frame}
 \item{"target.mut.sigs"}{a character vector}
 \item{"sequencing.artifact.mut.sigs"}{a charcter vector}
 \item{"config.obj"}{\code{\link{ParseConfigFile}}}
}}

\item{string}{= binary representation of parameter values}
}
\value{
A numeric value corresponding to the optimization objective value
}
\description{
Objective function for FIREVAT optimization
}
\keyword{internal}
