% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_optimization.R
\name{GAOptimizationObjFnHelper}
\alias{GAOptimizationObjFnHelper}
\title{GAOptimizationObjFnHelper}
\usage{
GAOptimizationObjFnHelper(params.x, data)
}
\arguments{
\item{params.x}{= representation of parameter values}

\item{data}{A list with the following data
\itemize{
 \item{"vcf.obj"}{\code{\link{ReadVCF}}}
 \item{"vcf.filter"}{a list}
 \item{"vcf.calling.method"}{string value}
 \item{"df.ref.mut.sig"}{a data.frame}
 \item{"target.mut.sigs"}{a character vector}
 \item{"sequencing.artifact.mut.sigs"}{a charcter vector}
 \item{"config.obj"}{\code{\link{ParseConfigFile}}}
}}
}
\value{
A list with the following elements
\itemize{
 \item{C.filtered}{numeric value}
 \item{A.filtered}{numeric value}
 \item{P.filtered}{numeric value}
 \item{C.artifact}{numeric value}
 \item{A.artifact}{numeric value}
 \item{P.artifact}{numeric value}
 \item{obj.val}{obj.val}
 \item{x}{x (numeric vector of GA parameter values)}
 \item{original mutations}{numeric value}
 \item{filtered.mutations}{numeric value}
 \item{artifact.mutations}{numeric value}
 \item{sequencing.artifact.signatures}{character vector (of signature names)}
 \item{sequencing.artifact.signatures.weights}{numeric vector}
}
}
\description{
Objective helper function for FIREVAT optimization
}
\keyword{internal}
