% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_config.R
\name{GenerateConfigObj}
\alias{GenerateConfigObj}
\title{Generate config.obj by checking vcf header}
\usage{
GenerateConfigObj(vcf.obj, save.config = TRUE,
  config.path = "../temp/FIREVAT_configure.json")
}
\arguments{
\item{vcf.obj}{A list from \code{\link{ReadVCF}}}

\item{save.config}{If true, save config.obj to config.path}

\item{config.path}{File path to write config.obj (json or yaml)}
}
\value{
config.obj
}
\description{
This function generate config.obj by checking vcf header.
Users should fill in the information needed in console.
In current version, only Integers & Float values can be used in
config.obj for running FIREVAT.
}
