% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_utils.R
\name{GetCOSMICMutSigs}
\alias{GetCOSMICMutSigs}
\title{GetCOSMICMutSigs}
\usage{
GetCOSMICMutSigs()
}
\value{
a data.frame of the COSMIC reference mutational signatures
}
\description{
Returns a data.frame of the COSMIC mutational signature reference file
from the data directory
}
