% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_utils.R
\name{GetCOSMICMutSigsEtiologiesColors}
\alias{GetCOSMICMutSigsEtiologiesColors}
\title{GetCOSMICMutSigsNames}
\usage{
GetCOSMICMutSigsEtiologiesColors()
}
\value{
data.frame with following columns: signature, group and color.
}
\description{
Returns all COSMIC mutational signature etiologies and colors
}
