% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_utils.R
\name{GetCOSMICMutSigsNames}
\alias{GetCOSMICMutSigsNames}
\title{GetCOSMICMutSigsNames}
\usage{
GetCOSMICMutSigsNames()
}
\value{
a character vector
}
\description{
Returns all COSMIC mutational signature names
}
