% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_report.R
\name{GetOptimizedSignatures}
\alias{GetOptimizedSignatures}
\title{GetOptimizedSignatures}
\usage{
GetOptimizedSignatures(data, mutations.type = "refined",
  signatures = "all")
}
\arguments{
\item{data}{A list of main data from \code{\link{RunFIREVAT}}}

\item{mutations.type}{A string for type of mutations ('refined' or 'artifact')}

\item{signatures}{A string ('all', 'target', 'artifact')}
}
\value{
A data.frame with the columns 'signature' and 'weight'
}
\description{
This function fetches the last row from the optimization iteration log
and returns the target and artifactual mutational signatures
for the type of mutations ('refined' or 'artifactual')
}
