% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_optimization.R
\name{GetParameterLowerUpperVector}
\alias{GetParameterLowerUpperVector}
\title{GetParameterLowerUpperVector}
\usage{
GetParameterLowerUpperVector(vcf.obj, config.obj, vcf.filter,
  multiplier = 100)
}
\arguments{
\item{vcf.obj}{A list from ReadVCF}

\item{config.obj}{A list from ParseConfigFile}

\item{vcf.filter}{A list from MakeMuTect2Filter}

\item{multiplier}{A multiplier for convert fraction to integer (default = 100)}
}
\value{
A list with the elements
\itemize{
 \item{lower.vector}{ A numeric vector. Each element is the minimum value of each parameter}
 \item{upper.vector}{ A numeric vector. Each element is the maximum value of each parameter}
 \item{vcf.obj}{ vcf.obj with updated data}
}
}
\description{
Return a lower/upper vector needed to conduct FIREVAT GA real-valued optimization.
}
\details{
vcf.obj$data: if max(vcf.obj$data[[param]]) < 1, then multiply multiplier to the vector
}
