% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_utils.R
\name{GetPCAWGMutSigs}
\alias{GetPCAWGMutSigs}
\title{GetPCAWGMutSigs}
\usage{
GetPCAWGMutSigs(sequencing.type = "wes")
}
\arguments{
\item{sequencing.type}{A string value.
It can be either 'wes' for whole-exome sequencing or 'wgs' for whole-genome sequencing}
}
\value{
a data.frame of the PCAWG mutatioanl signatures
}
\description{
Returns the PCAWG mutational signatures data
}
