% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_utils.R
\name{GetPCAWGMutSigsEtiologiesColors}
\alias{GetPCAWGMutSigsEtiologiesColors}
\title{GetPCAWGMutSigsEtiologiesColors}
\usage{
GetPCAWGMutSigsEtiologiesColors()
}
\value{
a data.frame with the columns 'signature', 'group', 'color'
}
\description{
Returns the PCAWG mutational signatures etiologies and colors
}
