% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_utils.R
\name{GetPCAWGMutSigsNames}
\alias{GetPCAWGMutSigsNames}
\title{GetPCAWGMutSigsNames}
\usage{
GetPCAWGMutSigsNames()
}
\value{
a character vector of the PCAWG mutational signatures names
}
\description{
Returns the PCAWG mutational signatures names
}
