% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_utils.R
\name{GetPCAWGPlatinumMutSigs}
\alias{GetPCAWGPlatinumMutSigs}
\title{GetPCAWGPlatinumMutSigs}
\usage{
GetPCAWGPlatinumMutSigs()
}
\value{
a data.frame of the PCAWG platinum mutatioanl signatures
}
\description{
Returns the PCAWG platinum mutational signatures data
}
