% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_utils.R
\name{GetPCAWGPlatinumMutSigsEtiologiesColors}
\alias{GetPCAWGPlatinumMutSigsEtiologiesColors}
\title{GetPCAWGPlatinumMutSigsEtiologiesColors}
\usage{
GetPCAWGPlatinumMutSigsEtiologiesColors()
}
\value{
a data.frame with the columns 'signature', 'group', 'color'
}
\description{
Returns the PCAWG platinum mutational signatures etiologies and colors
}
