% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_utils.R
\name{GetPCAWGPlatinumMutSigsNames}
\alias{GetPCAWGPlatinumMutSigsNames}
\title{GetPCAWGPlatinumMutSigsNames}
\usage{
GetPCAWGPlatinumMutSigsNames()
}
\value{
a character vector of the PCAWG platinum mutational signatures names
}
\description{
Returns the PCAWG platinum mutational signatures names
}
