% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_utils.R
\name{GetSeedForVCF}
\alias{GetSeedForVCF}
\title{GetSeedForVCF}
\usage{
GetSeedForVCF(vcf.file)
}
\arguments{
\item{vcf.file}{(full path of a .vcf file)}
}
\value{
an integer value
}
\description{
Returns a seed integer based on VCF file size
}
\details{
Returns the same seed integer for the same VCF file (based on file size)
}
