% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_vcf.R
\name{GetVCFValues}
\alias{GetVCFValues}
\title{GetVCFValues}
\usage{
GetVCFValues(vcf.obj, vcf.filter)
}
\arguments{
\item{vcf.obj}{A list of vcf data}

\item{vcf.filter}{A list of vcf filtering information}
}
\value{
A list with filtering parameter values
}
\description{
Get values of filtering parameters from vcf.obj
}
\keyword{internal}
