% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_vcf.R
\name{InitializeVCF}
\alias{InitializeVCF}
\title{InitializeVCF}
\usage{
InitializeVCF(vcf.obj, config.obj, verbose = TRUE)
}
\arguments{
\item{vcf.obj}{A list from ReadVCF}

\item{config.obj}{A list from ParseConfigFile}

\item{verbose}{If true, provides process detail}
}
\value{
A list with the following elements
\itemize{
 \item{vcf.obj.filtered}{vcf.obj (high-quality vcf)}
 \item{vcf.obj.artifact}{vcf.obj (low-quality vcf)}
}
}
\description{
Initialize VCF with FIREVAT config file
This functions selects point mutations and
appends filter values to vcf.obj$data
}
