% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_filter.R
\name{MakeFilter}
\alias{MakeFilter}
\title{MakeFilter}
\usage{
MakeFilter(config.obj)
}
\arguments{
\item{config.obj}{A list from ParseConfigFile
(any filter with "use_in_filter" value declared as FALSE is not considered)}
}
\value{
A list with the filter parameters
}
\description{
Creates a vcf filter from config.obj
}
