% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_filter.R
\name{MakeFilterFromConfig}
\alias{MakeFilterFromConfig}
\title{MakeFilterFromConfig}
\usage{
MakeFilterFromConfig(config.obj)
}
\arguments{
\item{config.obj}{A list from ParseConfigFile
(any filter with FALSE "use_in_filter" value is not considered)}
}
\value{
A list with the filter parameters
}
\description{
Creates a vcf filter from config.obj
}
