% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_mutalisk.R
\name{MutaliskParseVCFObj}
\alias{MutaliskParseVCFObj}
\title{MutaliskParseVCFObj}
\usage{
MutaliskParseVCFObj(vcf.obj)
}
\arguments{
\item{vcf.obj}{A list from ReadVCF}
}
\value{
A data.frame
}
\description{
Parses a vcf.obj and prepares it to run Mutalisk.
}
