% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_mutalisk.R
\name{MutaliskSigIndicesToNames}
\alias{MutaliskSigIndicesToNames}
\title{MutaliskSigIndicesToNames}
\usage{
MutaliskSigIndicesToNames(df.ref.mut.sigs, target.mut.sigs.indices)
}
\arguments{
\item{df.ref.mut.sigs}{A data.frame of reference mutational signatures}

\item{target.mut.sigs.indices}{An integer vector of target mutational signatures indices
(the first PCAWG signature, SBS1, has a value of 1)}
}
\value{
A character vector of target mutational signatures names
}
\description{
Converts target mutational signatures indices to the corresponding target
mutational signatures integer vector corresponding to the indices in df.ref.mut.sigs
}
\keyword{internal}
