% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_mutalisk.R
\name{MutaliskSigNamesToIndices}
\alias{MutaliskSigNamesToIndices}
\title{MutaliskSigNamesToIndices}
\usage{
MutaliskSigNamesToIndices(df.ref.mut.sigs, target.mut.sigs)
}
\arguments{
\item{df.ref.mut.sigs}{A data.frame of reference mutational signatures}

\item{target.mut.sigs}{A character vector of target mutational signatures names}
}
\value{
An integer vector of target mutational signatures indices
(the first target mutational signature (e.g. 'SBS1') gets assigned a value of 1)
}
\description{
Converts target.mut.sigs to an integer vector corresponding to the indices in df.ref.mut.sigs
}
\keyword{internal}
