% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_mutational_patterns.R
\name{MutPatParseRefMutSigs}
\alias{MutPatParseRefMutSigs}
\title{MutPatParseRefMutSigs}
\usage{
MutPatParseRefMutSigs(df.ref.mut.sigs, target.mut.sigs,
  signature.start.column.index = 4,
  mutation.type.header = "SomaticMutationType")
}
\arguments{
\item{df.ref.mut.sigs}{A data.frame of reference mutational signatures}

\item{target.mut.sigs}{A character vector of target mutational signatures names}

\item{signature.start.column.index}{= An integer value (e.g. column index corresponding to 'SBS1')}

\item{mutation.type.header}{= A string value
(name of header corresponding to column containing 'A[C>A]A' data))}
}
\value{
A data.frame of the format deconstructSigs::signatures.cosmic
}
\description{
Parses a df.ref.mut.sigs and prepares it to run Mutational Patterns.
}
