% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_optimization.R
\name{ParameterToBits}
\alias{ParameterToBits}
\title{ParameterToBits}
\usage{
ParameterToBits(vcf.obj, config.obj, vcf.filter, multiplier = 100)
}
\arguments{
\item{vcf.obj}{A list from ReadVCF}

\item{config.obj}{A list from ParseConfigFile}

\item{vcf.filter}{A list from MakeMuTect2Filter}

\item{multiplier}{A multiplier for convert fraction to integer (default = 100)}
}
\value{
A list with the elements
\itemize{
 \item{params.bit.len}{A numeric vector. Each element is the bit length of each parameter value}
 \item{vcf.obj}{A vcf.obj (\code{\link{ReadVCF}}) with updated data}
}
}
\description{
Calculate the number of bits needed to conduct FIREVAT GA binary optimization.
}
\details{
vcf.obj$data: if max(vcf.obj$data[[param]]) < 1, then multiply multiplier to the vector
}
