% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_config.R
\name{ParseConfigFile}
\alias{ParseConfigFile}
\title{ParseConfigFile}
\usage{
ParseConfigFile(config.path, verbose = TRUE)
}
\arguments{
\item{config.path}{A string for config file path}

\item{verbose}{If true, provides process detail}
}
\value{
config.obj: list of parameters
}
\description{
This function returns config.obj from JSON or YAML config file.
- Check if the config file is in JSON format or YAML format
- Return config.obj
}
\examples{
\dontrun{
ParseConfigFile("example.variant.caller.json")
ParseConfigFile("example.variant.caller.json", verbose=False)
}
}
