% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_constants.R
\docType{data}
\name{PCAWG.Platinum.All.Technology.Related.Artifact.Signatures}
\alias{PCAWG.Platinum.All.Technology.Related.Artifact.Signatures}
\title{Constant}
\format{An object of class \code{character} of length 9.}
\usage{
PCAWG.Platinum.All.Technology.Related.Artifact.Signatures
}
\description{
PCAWG mutational signatures reported to be associated with sequencing artifacts
}
\keyword{datasets}
