% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_constants.R
\docType{data}
\name{PCAWG.Possible.Sequencing.Artifact.Signatures}
\alias{PCAWG.Possible.Sequencing.Artifact.Signatures}
\title{Constant}
\format{An object of class \code{character} of length 17.}
\usage{
PCAWG.Possible.Sequencing.Artifact.Signatures
}
\description{
PCAWG mutational signatures reported to be associated with sequencing artifacts
}
\keyword{datasets}
