% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_constants.R
\docType{data}
\name{PCAWG.Target.Mutational.Signatures}
\alias{PCAWG.Target.Mutational.Signatures}
\title{Constant}
\format{An object of class \code{character} of length 47.}
\usage{
PCAWG.Target.Mutational.Signatures
}
\description{
PCAWG target mutational signatures reported to be unrelated to sequencing artifacts
}
\keyword{datasets}
