% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_strand_bias.R
\name{PerformStrandBiasAnalysis}
\alias{PerformStrandBiasAnalysis}
\title{PerformStrandBiasAnalysis}
\usage{
PerformStrandBiasAnalysis(vcf.obj, ref.forward.strand.var,
  ref.reverse.strand.var, alt.forward.strand.var, alt.reverse.strand.var,
  perform.fdr.correction = TRUE, fdr.correction.method = "BH")
}
\arguments{
\item{vcf.obj}{\code{\link{ReadVCF}}}

\item{ref.forward.strand.var}{A string value.}

\item{ref.reverse.strand.var}{A string value.}

\item{alt.forward.strand.var}{A string value.}

\item{alt.reverse.strand.var}{A string value.}

\item{perform.fdr.correction}{If TRUE, then performs false discovery rate correction}

\item{fdr.correction.method}{A string value. FDR correction method (Refer to p.adjust() function)}
}
\value{
An updated vcf.obj
}
\description{
Performs strand bias analysis
}
