% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_plot.R
\name{PlotMutaliskResults}
\alias{PlotMutaliskResults}
\title{PlotMutaliskResults}
\usage{
PlotMutaliskResults(mutalisk.results, signatures,
  df.ref.sigs.groups.colors, trinuc.max.y, trinuc.min.y, mut.type.max.y,
  title, font.size.small = 8, font.size.med = 14)
}
\arguments{
\item{mutalisk.results}{A list obtained from \code{\link{RunMutalisk}}}

\item{signatures}{A character vector of mutational signatures names}

\item{df.ref.sigs.groups.colors}{A data.frame with signature groups and colors}

\item{trinuc.max.y}{A numeric value (maximum y-axis value)}

\item{trinuc.min.y}{A numeric value (minimum y-axis value)}

\item{mut.type.max.y}{A numeric value}

\item{title}{A string value}

\item{font.size.small}{A numeric value}

\item{font.size.med}{A numeric value}
}
\value{
A ggplot object
}
\description{
Plots Mutalisk results
}
\examples{
\dontrun{
  df.ref.mut.sigs <- GetPCAWGMutSigs()
  target.mut.sigs <- GetPCAWGMutSigsNames()
  vcf.obj <- ReadVCF(vcf.file = "../data/sample/P-233-CT.final.vcf")
  mutalisk.results <- RunMutalisk(vcf.obj = vcf.obj,
                                  df.ref.mut.sigs = df.ref.mut.sigs,
                                  target.mut.sigs = target.mut.sigs)
  p <- PlotMutaliskResults(mutalisk.results = mutalisk.results)
  print(p)
}
}
