% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_plot.R
\name{PlotMutationTypes}
\alias{PlotMutationTypes}
\title{PlotMutationTypes}
\usage{
PlotMutationTypes(mutation.types = c("C>A", "C>G", "C>T", "T>A", "T>C",
  "T>G"), mutation.types.values, mutation.types.colors, max.y.val, title,
  convert.to.percentage = T, show.legend = T, font.size.small = 8,
  font.size.med = 14, plot.margin = unit(c(0.5, 0.5, 0.5, 0.5), "cm"))
}
\arguments{
\item{mutation.types}{Mutation types; Default = c("C>A", "C>G", "C>T", "T>A", "T>C", "T>G")}

\item{mutation.types.values}{Mutation count for each mutation type}

\item{mutation.types.colors}{A color vector for indicating mutation types}

\item{max.y.val}{y axis maximum value}

\item{title}{Plot title}

\item{convert.to.percentage}{if True convert y values to percentage (x 100); Default = T}

\item{show.legend}{If True, show legend; Default = T}

\item{font.size.small}{Small font size; Default = 8}

\item{font.size.med}{Medium font size; Default = 14}

\item{plot.margin}{Margin vector for drawing plot; Default = unit(c(0.5, 0.5, 0.5, 0.5), "cm"))}
}
\value{
A ggplot object
}
\description{
Plots a horizontal barplot of mutation types
}
\examples{
\dontrun{
p <- PlotMutationTypes(mutation.types = c("C>A", "C>G", "C>T", "T>A", "T>C", "T>G"),
                  mutation.types.values = c(0.3, 0.3, 0.1, 0.1, 0.1, 0.1),
                  mutation.types.colors = TriNuc.Mutation.Type.Hex.Colors,
                  max.y.val = 0.5,
                  convert.to.percentage = T,
                  show.legend = T,
                  font.size.small = 8,
                  font.size.med = 14,
                  plot.margin = unit(c(0.5, 0.5, 0.5, 0.5), "cm"))
print(p)
}
}
