% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_plot.R
\name{PlotOptimizationIterations}
\alias{PlotOptimizationIterations}
\title{PlotOptimizationIterations}
\usage{
PlotOptimizationIterations(df, columns.to.plot, x.axis.var, x.axis.title,
  x.min, x.max, save.file, title, y.axis.title = "", y.max = 1,
  point.size = 1, connect.dots = T, plot.legend = T,
  legend.ncol = 1, font.size.med = 14, font.size.large = 16,
  plot.margin = unit(c(0.5, 0.5, 0.5, 0.5), "cm"))
}
\arguments{
\item{df}{A data.frame (from reading "FIREVAT_Optimization_Logs.tsv")}

\item{columns.to.plot}{A character vector (of column names to plot)}

\item{x.axis.var}{x axis variable}

\item{x.axis.title}{x axis title}

\item{x.min}{x axis minimum value}

\item{x.max}{x axis maximum value}

\item{save.file}{Filename (including full path) to which the plot will be saved}

\item{title}{Plot title}

\item{y.axis.title}{y axis title; Default = ""}

\item{y.max}{y axis maximum value; Default = 1}

\item{point.size}{Point size; Default = 1}

\item{connect.dots}{If True draws dots for each iteration; Default = True}

\item{plot.legend}{If True write legend of plot; Default = T}

\item{legend.ncol}{legend.n Default = 1}

\item{font.size.med}{Medium font size; Default = 14}

\item{font.size.large}{Large font size; Default = 16}

\item{plot.margin}{Margin vector for plot; Default = unit(c(0.5, 0.5, 0.5, 0.5), "cm"))}
}
\value{
A ggplot object
}
\description{
Plots multiple scatter plots into one figure
}
