% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_plot.R
\name{PlotSignaturesContProbs}
\alias{PlotSignaturesContProbs}
\title{PlotSignaturesContProbs}
\usage{
PlotSignaturesContProbs(df.identified.mut.sigs, df.ref.sigs.groups.colors,
  title, convert.to.percentage = T, font.size.small = 8,
  font.size.med = 14, plot.margin = unit(c(0.5, 0.5, 0.5, 0.5), "cm"))
}
\arguments{
\item{df.identified.mut.sigs}{A data.frame of identified mutational signatures}

\item{df.ref.sigs.groups.colors}{A data.frame with 'signature', 'group', and 'color' columns}

\item{title}{Plot title}

\item{convert.to.percentage}{If true, convert y values to percentage (x 100); Default = T,}

\item{font.size.small}{Small font size; Default = 8,}

\item{font.size.med}{Medium font size; Default = 14,}

\item{plot.margin}{Margin vector for drawing plot; Default = unit(c(0.5, 0.5, 0.5, 0.5), "cm"))}
}
\value{
A ggplot object
}
\description{
Plots a horizontal barplot of identified mutational signatures
}
\examples{
\dontrun{
 g <- PlotSignaturesContProbs(sigs = c(mutalisk.results$identified.mut.sigs),
 sigs.probs = c(mutalisk.results$identified.mut.sigs.probs),
 df.ref.sigs.groups.colors = GetPCAWGMutSigsEtiologiesColors())
 print(g)
}
}
