% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_plot.R
\name{PlotTable}
\alias{PlotTable}
\title{PlotTable}
\usage{
PlotTable(df, padding = 20, font.size = 14)
}
\arguments{
\item{df}{= A data.frame where the first column is header and the second column is data value}

\item{padding}{Padding size; Default = 20}

\item{font.size}{Font size; Default = 14}
}
\value{
A plot
}
\description{
Plots basic statistics table
}
