% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_plot.R
\name{PlotTriNucSpectrum}
\alias{PlotTriNucSpectrum}
\title{PlotTriNucSpectrum}
\usage{
PlotTriNucSpectrum(sub.types, spectrum, max.y.val, min.y.val, y.axis.title,
  draw.top.strip = T, draw.x.axis.labels = T, draw.y.axis.labels = T,
  draw.y.axis.title = T, font.size.small = 8, font.size.med = 14,
  plot.margin.top = 0.5, plot.margin.bottom = 0.5,
  plot.margin.left = 0.5, plot.margin.right = 0.5, title)
}
\arguments{
\item{sub.types}{A character vector (types of 96 trinucleotide substitutions)}

\item{spectrum}{A numeric vector (96 elements)}

\item{max.y.val}{y axis maximum value}

\item{min.y.val}{y axis minimum value}

\item{y.axis.title}{y axis title}

\item{draw.top.strip}{If True then draws top strip; Default = T}

\item{draw.x.axis.labels}{If True then draws x axis labels; Default = T}

\item{draw.y.axis.labels}{If True then draws y axis labels; Default = T}

\item{draw.y.axis.title}{If True then draws y axis title; Default = T}

\item{font.size.small}{Small font size; Default = 8}

\item{font.size.med}{Medium font size; Default = 14}

\item{plot.margin.top}{Top margin; Default = 0.5}

\item{plot.margin.bottom}{Bottom margin; Default = 0.5}

\item{plot.margin.left}{Left margin; Default = 0.5}

\item{plot.margin.right}{Right margin; Default = 0.5}

\item{title}{Plot title}
}
\value{
A ggplot object
}
\description{
Plots the spectrum of 96 trinucleotide distribution
(C>A, C>G, C>T, T>A, T>C, T>G)
Please note that this function assumes that both sub.types and spectrum
are sorted in the following order: C>A, C>G, C>T, T>A, T>C, T>G
}
