% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_plot.R
\name{PlotVCFStatsBoxPlots}
\alias{PlotVCFStatsBoxPlots}
\title{PlotVCFStatsBoxPlots}
\usage{
PlotVCFStatsBoxPlots(original.vcf.stat.values, refined.vcf.stat.values,
  artifact.vcf.stat.values, xlab, axis.font.size = 10,
  label.font.size = 10, title.font.size = 12)
}
\arguments{
\item{original.vcf.stat.values}{A numeric vector corresponding to the original vcf.obj values of single filter parameter}

\item{refined.vcf.stat.values}{A numeric vector corresponding to the refined vcf.obj values of single filter parameter}

\item{artifact.vcf.stat.values}{A numeric vector corresponding to the artifact vcf.obj values of single filter parameter}

\item{xlab}{A string value (x-axis label)}

\item{axis.font.size}{An integer value (axis font size)}

\item{label.font.size}{An integer value (label font size)}

\item{title.font.size}{An integer value (title font size)}
}
\value{
A ggboxplot
}
\description{
Plots multiple (original, refined, artifact vcf) boxplots for single filter parameter
}
