% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_plot.R
\name{PlotVCFStatsHistograms}
\alias{PlotVCFStatsHistograms}
\title{PlotVCFStatsHistograms}
\usage{
PlotVCFStatsHistograms(plot.values, x.axis.labels, stat.y.max.vals,
  stat.x.max.vals, sample.id, save.file, title, cutoff.values,
  plot.boxplot = F, plot.cutoff.line.color = "#D4012E",
  plot.cutoff.value.lines = F, bin.width = 1, ncol = 4, nrow = 3,
  font.size.med = 10, font.size.large = 12, plot.margin = unit(c(0.5,
  0.5, 0.5, 0.5), "cm"))
}
\arguments{
\item{plot.values}{A list of multiple numeric vectors}

\item{x.axis.labels}{A character vector of x axis labels}

\item{stat.y.max.vals}{A numeric vector of max y-axis values}

\item{stat.x.max.vals}{A numeric vector of max x-axis values}

\item{sample.id}{A string value of sample ID}

\item{save.file}{A string value of file to which the resulting plot will be saved}

\item{title}{A string value of plot title}

\item{cutoff.values}{A numeric vector of cutoff values}

\item{plot.boxplot}{A boolean value (default = False)}

\item{plot.cutoff.line.color}{A hex string value (default = "#D4012E")}

\item{plot.cutoff.value.lines}{A boolean value (default = False)}

\item{bin.width}{An integer value (default = 1; histogram bin width)}

\item{ncol}{An integer value (default = 4; ggarrange ncol)}

\item{nrow}{An integer value (default = 3; ggarrange nrow)}

\item{font.size.med}{An integer value (default = 10)}

\item{font.size.large}{An integer value (default = 12)}

\item{plot.margin}{A list (default = unit(c(0.5, 0.5, 0.5, 0.5), "cm"))}
}
\value{
A list with the following elements
\itemize{
 \item{f}{ = A ggarrange object}
 \item{graphs}{ = A list of length 3; each element is a ggplot histogram}
}
}
\description{
Plots multiple VCF stats histograms into one figure
}
