% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_annotation.R
\name{PrepareAnnotationDB}
\alias{PrepareAnnotationDB}
\title{PrepareAnnotationDB}
\usage{
PrepareAnnotationDB(annotation.vcf.obj)
}
\arguments{
\item{annotation.vcf.obj}{vcf.obj of COSMIC or ClinVar vcf file}
}
\value{
A data.frame of annotation.vcf.obj
}
\description{
Prepares df.genes.of.interest from a vcf.obj (\code{\link{ReadVCF}})
of COSMIC or ClinVar vcf for \code{\link{AnnotateVCFObj}}
}
