% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_report.R
\name{PrepareArtifactAnnotationTable}
\alias{PrepareArtifactAnnotationTable}
\title{PrepareArtifactAnnotationTable}
\usage{
PrepareArtifactAnnotationTable(data)
}
\arguments{
\item{data}{A list of elements returned from \code{\link{RunFIREVAT}}}
}
\value{
A data.frame
}
\description{
Prepares artifactual mutations annotation (filtered, queried) table
}
