% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_report.R
\name{PrepareArtifactStrandBiasTable}
\alias{PrepareArtifactStrandBiasTable}
\title{PrepareArtifactStrandBiasTable}
\usage{
PrepareArtifactStrandBiasTable(data)
}
\arguments{
\item{data}{A list of elements returned from \code{\link{RunFIREVAT}}}
}
\value{
A data.frame
}
\description{
Prepares artifactual mutations strand biased variants table
}
