% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_report.R
\name{PrepareArtifactualMutsOptimizationIterationsPlot}
\alias{PrepareArtifactualMutsOptimizationIterationsPlot}
\title{PrepareArtifactualMutsOptimizationIterationsPlot}
\usage{
PrepareArtifactualMutsOptimizationIterationsPlot(data)
}
\arguments{
\item{data}{A list of elements returned from \code{\link{RunFIREVAT}}}
}
\value{
A ggplot object
}
\description{
Prepares artifactual mutations optimization iterations plot
}
