% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_report.R
\name{PrepareFilterCutoffsTable}
\alias{PrepareFilterCutoffsTable}
\title{PrepareFilterCutoffsTable}
\usage{
PrepareFilterCutoffsTable(data)
}
\arguments{
\item{data}{A list of elements returned from \code{\link{RunFIREVAT}}}
}
\value{
A data.frame
}
\description{
Prepares filter cutoffs table for reporting
}
