% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_report.R
\name{PrepareGeneticAlgorithmParametersTable}
\alias{PrepareGeneticAlgorithmParametersTable}
\title{PrepareGeneticAlgorithmParametersTable}
\usage{
PrepareGeneticAlgorithmParametersTable(data)
}
\arguments{
\item{data}{A list of elements returned from \code{\link{RunFIREVAT}}}
}
\value{
A data.frame
}
\description{
Prepares Genetic Algorithm parameters table
}
