% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_report.R
\name{PrepareIdentifiedSignaturesPlot}
\alias{PrepareIdentifiedSignaturesPlot}
\title{PrepareIdentifiedSignaturesPlot}
\usage{
PrepareIdentifiedSignaturesPlot(data)
}
\arguments{
\item{data}{A list of elements returned from \code{\link{RunFIREVAT}}}
}
\value{
A ggarrange object
}
\description{
Prepares identified signatures plot for reporting
}
