% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_report.R
\name{PrepareMLEReconstructedSpectrumsPlot}
\alias{PrepareMLEReconstructedSpectrumsPlot}
\title{PrepareMLEReconstructedSpectrumsPlot}
\usage{
PrepareMLEReconstructedSpectrumsPlot(data)
}
\arguments{
\item{data}{A list of elements returned from \code{\link{RunFIREVAT}}}
}
\value{
A ggarrange object
}
\description{
Prepares MLE reconstructed spectrums plot
}
