% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_report.R
\name{PrepareNucleotideSubstitutionTypesPlot}
\alias{PrepareNucleotideSubstitutionTypesPlot}
\title{PrepareNucleotideSubstitutionTypesPlot}
\usage{
PrepareNucleotideSubstitutionTypesPlot(data)
}
\arguments{
\item{data}{A list of elements returned from \code{\link{RunFIREVAT}}}
}
\value{
A ggarrange object
}
\description{
Prepares nucleotide substitution types plot
}
