% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_report.R
\name{PrepareOptimizationResultsTable}
\alias{PrepareOptimizationResultsTable}
\title{PrepareOptimizationResultsTable}
\usage{
PrepareOptimizationResultsTable(data)
}
\arguments{
\item{data}{A list of elements returned from \code{\link{RunFIREVAT}}}
}
\value{
A data.frame
}
\description{
Prepares optimization results table
}
