% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_report.R
\name{PrepareOptimizedVCFStatisticsPlot}
\alias{PrepareOptimizedVCFStatisticsPlot}
\title{PrepareOptimizedVCFStatisticsPlot}
\usage{
PrepareOptimizedVCFStatisticsPlot(data)
}
\arguments{
\item{data}{A list of elements returned from \code{\link{RunFIREVAT}}}
}
\value{
A ggarrange object
}
\description{
Prepares optimized VCF statistics plot
}
