% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_report.R
\name{PrepareRefinedMutsOptimizationIterationsPlot}
\alias{PrepareRefinedMutsOptimizationIterationsPlot}
\title{PrepareRefinedMutsOptimizationIterationsPlot}
\usage{
PrepareRefinedMutsOptimizationIterationsPlot(data)
}
\arguments{
\item{data}{A list of elements returned from \code{\link{RunFIREVAT}}}
}
\value{
A ggplot object
}
\description{
Prepares refined mutations optimization iterations plot
}
