% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_report.R
\name{PrepareRefinedStrandBiasTable}
\alias{PrepareRefinedStrandBiasTable}
\title{PrepareRefinedStrandBiasTable}
\usage{
PrepareRefinedStrandBiasTable(data)
}
\arguments{
\item{data}{A list of elements returned from \code{\link{RunFIREVAT}}}
}
\value{
A data.frame
}
\description{
Prepares refined mutations strand biased variants table
}
