% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_report.R
\name{PrepareReportDatafromConfig}
\alias{PrepareReportDatafromConfig}
\title{PrepareReportDatafromConfig}
\usage{
PrepareReportDatafromConfig(x.solution.decimal, data)
}
\arguments{
\item{x.solution.decimal}{A numeric vector}

\item{data}{A list of main data from \code{\link{RunFIREVAT}}}
}
\value{
A list with following items:
\itemize{
 \item{df.parameters}{A data.frame of FIREVAT Genetic Algorithm (GA) Parameters}
 \item{df.filter.cutoffs}{A data.frame of variant filtering cutoffs}
 \item{df.results}{A data.frame of objective values}
 \item{df.mutalisk.results}{A data.frame of mutational signature analysis results from Mutalisk}
}
}
\description{
Prepares report data (table data)
}
