% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_report.R
\name{PrepareReportPlotsfromConfig}
\alias{PrepareReportPlotsfromConfig}
\title{PrepareReportPlotsfromConfig}
\usage{
PrepareReportPlotsfromConfig(x.solution.decimal, data)
}
\arguments{
\item{x.solution.decimal}{A numeric vector}

\item{data}{A list of main data from \code{\link{RunFIREVAT}}}
}
\value{
A list with following items:
\itemize{
 \item{optimization.iter.plot}{A plot for the trend of objective value during iteraction}
 \item{refined.muts.seq.art.iter.plot}{A plot for showing sequencing artifact weights in refined variants}
 \item{artifactual.muts.seq.art.iter.plot}{A plot for showing sequencing artifact types in artifactual variants}
 \item{raw.muts.mutalisk.plots}{A plot for mutalisk results of original variants}
 \item{refined.muts.mutalisk.plots}{A plot for mutalisk results of refined variants}
 \item{artifactual.muts.mutalisk.plots}{A plot for mutalisk results of artifactual variants}
 \item{original.vcf.stats.plots}{A histogram for original VCF stats}
 \item{refined.vcf.stats.plots}{A histogram for refined VCF stats}
 \item{artifact.vcf.stats.plots}{A histogram for artifactual VCF stats}
 \item{vcf.stats.boxplots}{A plot for showing comparison results among VCF stats of original, refined and artifactual VCFs}
}
}
\description{
Prepares report plots
}
