% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_report.R
\name{PrepareTrinucleotideSpectrumsTable}
\alias{PrepareTrinucleotideSpectrumsTable}
\title{PrepareTrinucleotideSpectrumsTable}
\usage{
PrepareTrinucleotideSpectrumsTable(data)
}
\arguments{
\item{data}{A list of elements returned from \code{\link{RunFIREVAT}}}
}
\value{
A data.frame
}
\description{
Prepares trinucleotide spectrums table
}
