% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_log.R
\name{PrintLog}
\alias{PrintLog}
\title{PrintLog}
\usage{
PrintLog(msg, type = "INFO")
}
\arguments{
\item{msg}{String value message to print along with log type and date}

\item{type}{String value that represents type of this message. 'INFO' by default.}
}
\description{
Prints log message
}
