% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_annotation.R
\name{QueryAnnotatedVCF}
\alias{QueryAnnotatedVCF}
\title{FilterAnnotatedVCF}
\usage{
QueryAnnotatedVCF(vcf.obj.annotated, filter.key.value.pairs,
  filter.condition = "AND")
}
\arguments{
\item{vcf.obj.annotated}{\code{\link{AnnotateVCFObj}}}

\item{filter.key.value.pairs}{A list with the key as the column name and value as the
filtering values. E.g. list("CLNSIG" = c("Pathogenic", "Pathogenic/Likely_pathogenic"))}

\item{filter.condition}{'AND' or 'OR'.}
}
\value{
A vcf.obj
}
\description{
Annotates a vcf.obj using df.variants.of.interest (from (\code{\link{PrepareAnnotationDB}})
}
