% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_report.R
\name{ReadOptimizationIterationReport}
\alias{ReadOptimizationIterationReport}
\title{ReadOptimizationIterationReport}
\usage{
ReadOptimizationIterationReport(data)
}
\arguments{
\item{data}{A list of elements returned from \code{\link{RunFIREVAT}}}
}
\value{
A data.frame of FIREVAT optimization logs
}
\description{
Read optimization iteration report
}
