% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_vcf.R
\name{ReadVCF}
\alias{ReadVCF}
\title{ReadVCF}
\usage{
ReadVCF(vcf.file, genome = "hg19")
}
\arguments{
\item{vcf.file}{(full path of a .vcf file)}

\item{genome}{('hg19' or 'hg38')}
}
\value{
A list with elements 'data', 'header', 'genome'
}
\description{
Reads a .vcf file
}
