% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_report.R
\name{ReportFIREVATResults}
\alias{ReportFIREVATResults}
\title{ReportFIREVATResults}
\usage{
ReportFIREVATResults(x.solution.decimal, data, verbose = T,
  report.format = "html")
}
\arguments{
\item{x.solution.decimal}{A numeric vector}

\item{data}{A list of main data from \code{\link{RunFIREVAT}}}

\item{verbose}{If true, provides process detail}

\item{report.format}{The format of FIREVAT report. In current version, we only support "html"}
}
\description{
Reports FIREVAT results in html format (generated from Rmd)
}
