% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_main.R
\name{RunFIREVAT}
\alias{RunFIREVAT}
\title{RunFIREVAT}
\usage{
RunFIREVAT(vcf.file, vcf.file.genome, config.file, df.ref.mut.sigs,
  target.mut.sigs, sequencing.artifact.mut.sigs, num.cores, output.dir,
  pop.size = 200, max.iter = 200, run = 50, pmutation = 0.25,
  verbose = TRUE)
}
\arguments{
\item{vcf.file}{String value corresponding to input .vcf file (full path)}

\item{vcf.file.genome}{Genome assembly of the input .vcf file}

\item{config.file}{String value corresponding to input configuration file (refer to ...)}

\item{df.ref.mut.sigs}{A data.frame of the reference mutational signatures}

\item{target.mut.sigs}{A character vector of the target mutational signatures (from reference mutational signatures)}

\item{sequencing.artifact.mut.sigs}{A character vector of the sequencing artifact mutational signatures (from reference mutational signatures)}

\item{num.cores}{Number of cores to allocate}

\item{output.dir}{String value of the desired output directory}

\item{pop.size}{Integer value of the Genetic Algorithm 'population size' parameter}

\item{max.iter}{Integer value of the Genetic Algorithm 'maximum iterations' parameter}

\item{run}{Integer value of the Genetic Algorithm 'run' parameter}

\item{pmutation}{Float value of the Genetic Algorithm 'mutation probability' parameter}

\item{verbose}{If true, provides process detail}
}
\value{
A list with the following elements
\itemize{
 \item{x.solution.decimal}{A numeric vector of optimized parameter values}
 \item{data}{}
}
}
\description{
Runs FIREVAT using configuration data
}
