% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_mutalisk.R
\name{RunMutalisk}
\alias{RunMutalisk}
\title{RunMutalisk}
\usage{
RunMutalisk(vcf.obj, df.ref.mut.sigs, target.mut.sigs,
  random.sampling.candidate.mut.sigs = c(), method = "random.sampling",
  n.sample = 20, n.iter = 10, verbose = TRUE)
}
\arguments{
\item{vcf.obj}{A list (from firevat_vcf::ReadVCF)}

\item{df.ref.mut.sigs}{A data.frame of reference mutational signatures}

\item{target.mut.sigs}{A character vector of target mutational signatures names to identify from}

\item{random.sampling.candidate.mut.sigs}{A character vector of mutational signatures names
that gets appended to the list of candidate mutational signatures so that these are
always considered.}

\item{method}{A string value (must be either 'random.sampling' or 'all').
The method 'random.sampling' samples (without replacement) 'n.sample' number of signatures
'n.iter' number of times and runs the candidate signatures one last time.
The method 'all' uses all target.mut.sigs}

\item{n.sample}{An integer value ('random.sampling' method parameter)
Number of signatures to choose for each iteration of random sampling).}

\item{n.iter}{An integer value ('random.sampling' method parameter).
Number of iterations to perform random sampling.}

\item{verbose}{If true, provides process details}
}
\value{
A list with the following elements
\itemize{
 \item{num.point.mutations}{An integer value - count of total point mutations}
 \item{sub.types}{A character vector of length 96}
 \item{sub.types.spectrum}{A numeric vector of length 96}
 \item{num.mut.sigs}{An integer value (count of unique mutational signatures identified)}
 \item{identified.mut.sigs}{A character vector where each element is a mutational signature identified}
 \item{identified.mut.sigs.probs}{A numeric vector where each element is the weight of mutational signature identified.
  The ordering follows identified.mut.sigs}
 \item{identified.mut.sigs.spectrum}{A numeric vector of length 96}
 \item{residuals}{A numeric vector of length 96}
 \item{rss}{A numeric value (residual sum of squares)}
 \item{cos.sim.score}{A numeric value (cosine similarity score between observed mutational spectrum and
  reconstructed mutational signatures)}
 \item{all.models.sigs}{A list where each element is a model; a model is a list of signatures identified)}
 \item{all.models.sigs.probs}{A list where each element is a model; a model is a list of contribution probabilities}
 \item{all.models.cos.sim.scores}{A list where each element is a model; a model is a list of cosine similarity socres}
}
}
\description{
Identifies mutational signatures using Mutalisk
}
