% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_mutalisk.R
\name{RunMutaliskHelper}
\alias{RunMutaliskHelper}
\title{RunMutaliskHelper}
\usage{
RunMutaliskHelper(vcf.trinucleotide.data, df.ref.mut.sigs, target.mut.sigs)
}
\arguments{
\item{vcf.trinucleotide.data}{A data.frame (from firevat_mutalisk::MutaliskParseVCFObj)}

\item{df.ref.mut.sigs}{A data.frame of reference mutational signatures}

\item{target.mut.sigs}{A character vector of target mutational signatures names}
}
\value{
A list with the following elements
\itemize{
 \item{num.point.mutations}{An integer value - count of total point mutations}
 \item{sub.types}{A character vector of length 96}
 \item{sub.types.spectrum}{A numeric vector of length 96}
 \item{num.mut.sigs}{An integer value (count of unique mutational signatures identified)}
 \item{identified.mut.sigs}{A character vector where each element is a mutational signature identified}
 \item{identified.mut.sigs.probs}{A numeric vector where each element is the weight of mutational signature identified.
  The ordering follows identified.mut.sigs}
 \item{identified.mut.sigs.spectrum}{A numeric vector of length 96}
 \item{residuals}{A numeric vector of length 96}
 \item{rss}{A numeric value (residual sum of squares)}
 \item{cos.sim.score}{A numeric value (cosine similarity score between observed mutational spectrum and
  reconstructed mutational signatures)}
 \item{all.models.sigs}{A list where each element is a model; a model is a list of signatures identified)}
 \item{all.models.sigs.probs}{A list where each element is a model; a model is a list of contribution probabilities}
 \item{all.models.cos.sim.scores}{A list where each element is a model; a model is a list of cosine similarity socres}
}
}
\description{
Helper function for RunMutalisk
}
