% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_mutational_patterns.R
\name{RunMutPat}
\alias{RunMutPat}
\title{RunMutPat}
\usage{
RunMutPat(vcf.obj, df.ref.mut.sigs, target.mut.sigs, verbose = TRUE)
}
\arguments{
\item{vcf.obj}{A list from ReadVCF}

\item{df.ref.mut.sigs}{A data.frame of reference mutational signatures}

\item{target.mut.sigs}{A character vector of target mutational signatures names}

\item{verbose}{If true, provides process details}
}
\value{
A list with the following elements
\itemize{
 \item{tumor.mutation.types.spectrum}{A numeric vector of length 96 - 'observed' spectrum}
 \item{identified.mutation.types.spectrum}{A numeric vector of length 96 - 'identified' spectrum}
 \item{residuals}{A numeric vector of length 96 - residuals}
 \item{mutation.types}{A character vector of length 96}
 \item{identified.mut.sigs}{A character vector where each element is a mutational signature identified}
 \item{identified.mut.sigs.contribution.weights}{A numeric vector where each element is the weight of mutational signature identified. The ordering follows identified.mut.sigs}
 \item{cosine.similarity.score}{A numeric value}
}
}
\description{
Identifies mutational signatures using Mutational Patterns
}
\examples{
\dontrun{
vcf.obj <- ReadVCF(vcf.file = "../data/sample/HNT-082-BT.final.call.vcf", genome = "hg19")
df.ref.mut.sigs <- GetPCAWGMutSigs()
target.mut.sigs <- GetPCAWGMutSigsNames()
RunMutPat(vcf.obj = vcf.obj,
df.ref.mut.sigs = df.ref.mut.sigs,
target.mut.sigs = target.mut.sigs)
}
}
\keyword{internal}
