% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_objective_functions.R
\name{Sig.Extraction.Obj.Fn}
\alias{Sig.Extraction.Obj.Fn}
\title{Sig.Extraction.Obj.Fn}
\usage{
Sig.Extraction.Obj.Fn(C.refined, A.refined, C.artifactual, A.artifactual)
}
\arguments{
\item{C.refined}{A numeric value between 0 and 1.}

\item{A.refined}{A numeric value between 0 and 1.}

\item{C.artifactual}{A numeric value between 0 and 1.}

\item{A.artifactual}{A numeric value between 0 and 1.}
}
\value{
A numeric value between 0 and 1.
}
\description{
Calculates the signature extraction objective value for FIREVAT GA optimization.
}
