% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_objective_functions.R
\name{Test.Obj.Fn.2}
\alias{Test.Obj.Fn.2}
\title{Test.Obj.Fn.2}
\usage{
Test.Obj.Fn.2(C.refined, A.refined, C.artifactual, A.artifactual)
}
\arguments{
\item{C.refined}{A numeric value between 0 and 1.}

\item{A.refined}{A numeric value between 0 and 1.}

\item{C.artifactual}{A numeric value between 0 and 1.}

\item{A.artifactual}{A numeric value between 0 and 1.}
}
\value{
A numeric value between 0 and 1.
}
\description{
Test objective function 2
}
