% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_config.R
\name{TransformValueType}
\alias{TransformValueType}
\title{TransformValueType}
\usage{
TransformValueType(curr.value, value_type)
}
\arguments{
\item{curr.value}{A value which need type conversion}

\item{value_type}{A character value}
}
\value{
curr.value. Type-converted value (integer or numeric)
}
\description{
This function checks the type of value and returns type-transformed value
}
\examples{
\dontrun{
TransformValueType("0.98", "double")
TransformValueType("3.0", "integer")
}
}
\keyword{internal}
