% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_constants.R
\docType{data}
\name{TriNuc.Mutation.Type.Hex.Colors}
\alias{TriNuc.Mutation.Type.Hex.Colors}
\title{Constant}
\format{An object of class \code{character} of length 6.}
\usage{
TriNuc.Mutation.Type.Hex.Colors
}
\description{
Hex codes for the mutation types (for plotting purposes)
}
\keyword{datasets}
