% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_utils.R
\name{WriteFIREVATResultsToTSV}
\alias{WriteFIREVATResultsToTSV}
\title{WriteFIREVATResultsToTSV}
\usage{
WriteFIREVATResultsToTSV(firevat.results)
}
\arguments{
\item{firevat.results}{List returned from RunFIREVAT}
}
\description{
Writes FIREVAT results to a csv file
}
