% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_vcf.R
\name{WriteVCF}
\alias{WriteVCF}
\title{WriteVCF}
\usage{
WriteVCF(vcf.obj, save.file)
}
\arguments{
\item{vcf.obj}{(from the function ReadVCF)}

\item{save.file}{(full path including filename)}
}
\description{
Writes a vcf.obj to a .vcf file
}
