% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_mutational_patterns.R
\name{MutPatParseVCFObj}
\alias{MutPatParseVCFObj}
\title{MutPatParseVCFObj}
\usage{
MutPatParseVCFObj(vcf.obj, bsg, sample.id = "sample")
}
\arguments{
\item{vcf.obj}{A list from ReadVCF}

\item{bsg}{A BSgenome object}

\item{sample.id}{A string value}
}
\value{
A data.frame with the column sample.id and
row names corresponding to 96 substitution types
}
\description{
Parses a vcf.obj and prepares it to run Mutational Patterns.
}
