% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_report.R
\name{PrepareIdentifiedSignaturesPlotHelper}
\alias{PrepareIdentifiedSignaturesPlotHelper}
\title{PrepareIdentifiedSignaturesPlotHelper}
\usage{
PrepareIdentifiedSignaturesPlotHelper(mutalisk.results, signatures, title,
  df.ref.mut.sigs.groups.colors)
}
\arguments{
\item{mutalisk.results}{A list of elements returned from \code{\link{RunMutalisk}}}

\item{signatures}{A character vector of signature names}

\item{title}{String value}

\item{df.ref.mut.sigs.groups.colors}{A data.frame with signature and color hex value}
}
\value{
A ggplot object
}
\description{
PrepareIdentifiedSignaturesPlot helper function
}
\keyword{internal}
