% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_report.R
\name{ReportFIREVATResults}
\alias{ReportFIREVATResults}
\title{ReportFIREVATResults}
\usage{
ReportFIREVATResults(data)
}
\arguments{
\item{data}{A list of main data from \code{\link{RunFIREVAT}}}
}
\value{
An updated data list
}
\description{
Reports FIREVAT results in html format (generated from Rmd)
}
