% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firevat_main.R
\name{RunFIREVAT}
\alias{RunFIREVAT}
\title{RunFIREVAT}
\usage{
RunFIREVAT(vcf.file, vcf.file.genome = "hg19",
  check.chromosome.name = TRUE, config.file,
  df.ref.mut.sigs = GetPCAWGMutSigs(),
  target.mut.sigs = GetPCAWGMutSigsNames(),
  df.ref.mut.sigs.groups.colors = GetPCAWGMutSigsEtiologiesColors(),
  sequencing.artifact.mut.sigs = PCAWG.All.Sequencing.Artifact.Signatures,
  num.cores = 2, output.dir, mode = "ga", init.artifact.stop = 0.05,
  objective.fn = Default.Obj.Fn, use.suggested.soln = TRUE,
  ga.type = "real-valued", ga.pop.size = 100, ga.max.iter = 100,
  ga.run = 50, ga.pmutation = 0.1, ga.preemptive.killing = FALSE,
  ga.seed = NULL, mutalisk = TRUE, mutalisk.method = "all",
  mutalisk.must.include.sigs = NULL,
  mutalisk.random.sampling.count = 20,
  mutalisk.random.sampling.max.iter = 10,
  perform.strand.bias.analysis = FALSE,
  filter.by.strand.bias.analysis = TRUE,
  filter.by.strand.bias.analysis.cutoff = 0.25,
  strand.bias.perform.fdr.correction = TRUE,
  strand.bias.fdr.correction.method = "BH",
  ref.forward.strand.var = NULL, ref.reverse.strand.var = NULL,
  alt.forward.strand.var = NULL, alt.reverse.strand.var = NULL,
  annotate = FALSE, df.annotation.db = NULL,
  annotated.columns.to.display = NULL,
  annotation.filter.key.value.pairs = NULL,
  annotation.filter.condition = "AND", write.vcf = TRUE,
  report = TRUE, save.rdata = TRUE, save.tsv = TRUE,
  report.format = "html", verbose = TRUE)
}
\arguments{
\item{vcf.file}{String value corresponding to input .vcf file. Please provide the full path.}

\item{vcf.file.genome}{Genome assembly of the input .vcf file. The genome should be supported by BSgenome.}

\item{check.chromosome.name}{Boolean value. If TRUE, FIREVAT checks chromosome names}

\item{config.file}{String value corresponding to input configuration file. For more details please refer to ...}

\item{df.ref.mut.sigs}{A data.frame of the reference mutational signatures}

\item{target.mut.sigs}{A character vector of the target mutational signatures from reference mutational signatures.}

\item{df.ref.mut.sigs.groups.colors}{A data.frame of the reference mutational signatures groups and colors}

\item{sequencing.artifact.mut.sigs}{A character vector of the sequencing artifact mutational signatures from reference mutational signatures.}

\item{num.cores}{Number of cores to allocate}

\item{output.dir}{String value of the desired output directory}

\item{mode}{String value. The value should be either 'ga' or 'manual'.}

\item{init.artifact.stop}{Numeric value  less than 1. If the sum of sequencing artifact weights in the user-specified original VCF file (i.e. vcf.file)
is less than or equal to this value then FIREVAT does not perform variant refinement.
Default value is 0.05. Note that this option does not apply if 'mode' is 'manual'.}

\item{objective.fn}{Objective value derivation function. Default: Default.Obj.Fn.}

\item{use.suggested.soln}{Boolean value. If TRUE, then FIREVAT passes the default values
of filter variables declared as 'use_in_filter' in the config file to the 'suggestions' parameter of
the Genetic Algorithm package. If FALSE, then FIREVAT supplies NULL to the GA package 'suggestions' parameter.
FIREVAT also computes baseline performance of each filter variable and uses fittest population from each variable
as a suggested solution.}

\item{ga.type}{String value. The value should be either 'binray' or 'real-valued'.}

\item{ga.pop.size}{Integer value of the Genetic Algorithm 'population size' parameter. Default: 100.
This value should be set based on the number of filter parameters. Recommendation: 40 per filter parameter.}

\item{ga.max.iter}{Integer value of the Genetic Algorithm 'maximum iterations' parameter. Default: 100.
This value should be set based on the number of filter parameters. Recommendation: same as 'ga.pop.size'.}

\item{ga.run}{Integer value of the Genetic Algorithm 'run' parameter. Default: 50.
This value should be set based on the 'ga.max.iter' parameter. Recommendation: 25 percent of 'ga.max.iter'.}

\item{ga.pmutation}{Float value of the Genetic Algorithm 'mutation probability' parameter. Default: 0.1.}

\item{ga.preemptive.killing}{If TRUE, then preemptively kills populations that yield greater sequencing artifact weights sum
compared to the original mutatational signatures analysis}

\item{ga.seed}{Integer value of the Genetic Algorithm 'seed' parameter. Default: NULL.}

\item{mutalisk}{If TRUE, confirm mutational signature analysis with Mutalisk. Default: TRUE.}

\item{mutalisk.method}{Mutalisk signature identification method. Default: 'random.sampling'.
The value can be either 'all' or 'random.sampling'.
'all' uses all target.mut.sigs to identify mutational signatures.
'random.sampling' randomly samples from target.mut.sigs to identify mutational signatures.}

\item{mutalisk.must.include.sigs}{Signatures that must be included in the Mutalisk signature identification
A character vector corresponding to the signature names.}

\item{mutalisk.random.sampling.count}{Mutalisk random sampling count. Default: 20.
The number of signatures to sample from target.mut.sigs}

\item{mutalisk.random.sampling.max.iter}{Mutalisk random sampling maximum iteration. Default: 10.
The number of times Mutalisk randomly samples from target.mut.sigs before determining the candidate signatures.}

\item{perform.strand.bias.analysis}{If TRUE, then performs strand bias analysis.}

\item{filter.by.strand.bias.analysis}{If TRUE, then filters out variants in refined vcf based on strand bias analysis results}

\item{filter.by.strand.bias.analysis.cutoff}{The p.value or q value cutoff for filtering out variants.}

\item{strand.bias.perform.fdr.correction}{If TRUE, then performs false discovery rate
correction for strand bias analysis.}

\item{strand.bias.fdr.correction.method}{A string value. Default value is 'BH'.
Refer to 'p.adjust()' function method.}

\item{ref.forward.strand.var}{A string value.}

\item{ref.reverse.strand.var}{A string value,}

\item{alt.forward.strand.var}{A string value,}

\item{alt.reverse.strand.var}{A string value,}

\item{annotate}{A boolean value. Default value is TRUE.}

\item{df.annotation.db}{A data.frame. Please refer to \code{\link{PrepareAnnotationDB}}}

\item{annotated.columns.to.display}{A character vector.}

\item{annotation.filter.key.value.pairs}{A list.}

\item{annotation.filter.condition}{'AND' or 'OR'.}

\item{write.vcf}{If TRUE, write original/refined/artifact vcfs. Default: TRUE.}

\item{report}{If TRUE, generate report. Default: TRUE.}

\item{save.rdata}{If TRUE, save rdata. Default: TRUE.}

\item{save.tsv}{If TRUE, save tsv. Default: TRUE.}

\item{report.format}{The format of FIREVAT report. We currently only support 'html'.}

\item{verbose}{If TRUE, provides process detail. Default: TRUE.}
}
\value{
A list with the following elements
\itemize{
 \item{f}{ = A ggarrange object}
 \item{graphs}{ = A list of length 3; each element is a ggplot histogram}
}
}
\description{
Runs FIREVAT using configuration data. Filters point mutations in the user-specified vcf file based on mutational signature
identification and outputs the refined and artifact vcf files as well as metadata related to the refinement process.
}
